+++ tests/Classes/Domain/Model/DataTest.php
@@ -1,54 +1,77 @@
 <?php
 /***************************************************************
-*  Copyright notice
-*
- *  (c) 2007-2014 Rene Nitzsche (rene@system25.de)
-*  All rights reserved
-*
-*  This script is part of the TYPO3 project. The TYPO3 project is
-*  free software; you can redistribute it and/or modify
-*  it under the terms of the GNU General Public License as published by
-*  the Free Software Foundation; either version 2 of the License, or
-*  (at your option) any later version.
-*
-*  The GNU General Public License can be found at
-*  http://www.gnu.org/copyleft/gpl.html.
-*
-*  This script is distributed in the hope that it will be useful,
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*  GNU General Public License for more details.
-*
-*  This copyright notice MUST APPEAR in all copies of the script!
-***************************************************************/
+ * Copyright notice
+ *
+ * (c) 2007-2016 Rene Nitzsche (rene@system25.de)
+ * All rights reserved
+ *
+ * This script is part of the TYPO3 project. The TYPO3 project is
+ * free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * The GNU General Public License can be found at
+ * http://www.gnu.org/copyleft/gpl.html.
+ *
+ * This script is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * This copyright notice MUST APPEAR in all copies of the script!
+ ***************************************************************/
 
 tx_rnbase::load('tx_rnbase_tests_BaseTestCase');
 tx_rnbase::load('Tx_Rnbase_Domain_Model_Data');
 
 /**
+ * Data model unit tests
  *
- * @package tx_rnbase
- * @subpackage tx_rnbase_tests
- * @author Michael Wagner <michael.wagner@dmk-ebusiness.de>
+ * @package TYPO3
+ * @subpackage Tx_Rnbase
+ * @author Michael Wagner
+ * @license http://www.gnu.org/licenses/lgpl.html
+ *          GNU Lesser General Public License, version 3 or later
  */
 class Tx_Rnbase_Domain_Model_DataTest
 	extends tx_rnbase_tests_BaseTestCase {
-
 	/**
-	 * test object with testdata
+	 * Test the getProperties method
 	 *
-	 * @return Tx_Rnbase_Domain_Model_Data
+	 * @return void
+	 *
+	 * @group unit
+	 * @test
 	 */
-	private function getModelInstance() {
-		$data = array(
-			'uid' => 50,
-			'first_name' => 'John',
-			'last_name' => 'Doe',
+	public function testGetPropertiesCallsProperty()
+	{
+		$model = $this->getModel(
+			null,
+			'Tx_Rnbase_Domain_Model_Base',
+			array('getProperty')
 		);
-		return Tx_Rnbase_Domain_Model_Data::getInstance($data);
+
+		$model
+			->expects(self::once())
+			->method('getProperty')
+			->with(self::equalTo(null))
+			->will(self::returnValue(array('uid' => 1)))
+		;
+
+		$data = $model->getProperties();
+
+		self::assertTrue(is_array($data));
+		self::assertCount(1, $data);
+		self::assertEquals(1, $data['uid']);
 	}
 
 	/**
+	 * Test Magic calls
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
 	public function testMagicCalls() {
@@ -76,6 +99,11 @@ class Tx_Rnbase_Domain_Model_DataTest
 	}
 
 	/**
+	 * Test record overloding for getters.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
 	public function testRecordOverloadingGet() {
@@ -86,18 +114,30 @@ class Tx_Rnbase_Domain_Model_DataTest
 	}
 
 	/**
+	 * Test if magic calls throw exception on unknown method.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 * @expectedException Exception
 	 * @expectedExceptionCode 1406625817
 	 */
-	public function testMagicCallThrowsException() {
+	public function testMagicCallThrowsException()
+	{
 		$this->getModelInstance()->methodDoesNotExist();
 	}
 
 	/**
+	 * Test getInstance with recursive data.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
-	public function testRecursiveInstance() {
+	public function testRecursiveInstance()
+	{
 		$data = array(
 			'gender' => 'm',
 			'name' => array(
@@ -116,18 +156,30 @@ class Tx_Rnbase_Domain_Model_DataTest
 	}
 
 	/**
+	 * Test dirty state.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
-	public function testIsDirtyOnGet() {
+	public function testIsDirtyOnGet()
+	{
 		$model = $this->getModelInstance();
 		$model->getFirstName('Jonny');
 		$this->assertFalse($model->isDirty());
 	}
 
 	/**
+	 * Test dirty state.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
-	public function testIsDirtyOnSet() {
+	public function testIsDirtyOnSet()
+	{
 		$model = $this->getModelInstance();
 		$model->setFirstName('Jonny');
 		// after set, the model has to be dirty
@@ -144,10 +196,17 @@ class Tx_Rnbase_Domain_Model_DataTest
 		);
 		$this->assertTrue($model->isDirty());
 	}
+
 	/**
+	 * Test dirty state.
+	 *
+	 * @return void
+	 *
+	 * @group unit
 	 * @test
 	 */
-	public function testIsDirtyOnUns() {
+	public function testIsDirtyOnUns()
+	{
 		$model = $this->getModelInstance();
 		// after unset an nonexisting value, the model has to be clean
 		$model->unsSomeNotExistingColumn();
@@ -156,8 +215,18 @@ class Tx_Rnbase_Domain_Model_DataTest
 		$model->unsFirstName();
 		$this->assertTrue($model->isDirty());
 	}
-}
 
-if (defined('TYPO3_MODE') && $TYPO3_CONF_VARS[TYPO3_MODE]['XCLASS']['ext/rn_base/tests/model/class.tx_rnbase_tests_model_Data_testcase.php']) {
-	include_once($TYPO3_CONF_VARS[TYPO3_MODE]['XCLASS']['ext/rn_base/tests/model/class.tx_rnbase_tests_model_Data_testcase.php']);
+	/**
+	 * Creates a test object
+	 *
+	 * @return Tx_Rnbase_Domain_Model_Data
+	 */
+	private function getModelInstance() {
+		$data = array(
+			'uid' => 50,
+			'first_name' => 'John',
+			'last_name' => 'Doe',
+		);
+		return Tx_Rnbase_Domain_Model_Data::getInstance($data);
+	}
 }